functions/requirements.txt
firebase-admin>=12.0.0
firebase-functions>=2.0.5
google-generativeai>=0.3.0
google-cloud-firestore>=2.21.0
google-cloud-storage>=3.10.0
pandas>=2.2.0
numpy>=1.26.3
functions-framework>=3.0.0
Flask>=3.0.0
python-dotenv>=1.0.0